+++ src/openapi/router.rs
@@ -3,7 +3,7 @@ use super::{
 	handler::{OpenapiHandler, SwaggerUiHandler},
 	operation::OperationDescription
 };
-use crate::{routing::*, EndpointWithSchema, OpenapiType, ResourceResultSchema, ResourceWithSchema};
+use crate::{routing::*, EndpointWithSchema, OpenapiType, ResourceWithSchema, ResponseSchema};
 use gotham::{hyper::Method, pipeline::chain::PipelineHandleChain, router::builder::*};
 use once_cell::sync::Lazy;
 use regex::{Captures, Regex};