+++ src/response/no_content.rs
@@ -1,9 +1,9 @@
-use super::{handle_error, ResourceResult};
+use super::{handle_error, IntoResponse};
 use crate::{IntoResponseError, Response};
 #[cfg(feature = "openapi")]
-use crate::{OpenapiSchema, OpenapiType, ResourceResultSchema};
-
+use crate::{OpenapiSchema, OpenapiType, ResponseSchema};
 use futures_util::{future, future::FutureExt};
+use gotham::hyper::header::{HeaderMap, HeaderValue, IntoHeaderName};
 #[cfg(feature = "openapi")]
 use gotham::hyper::StatusCode;
 use mime::Mime;
@@ -31,22 +31,36 @@ fn read_all() {
 # }
 ```
 */
-#[derive(Clone, Copy, Debug, Default)]
-pub struct NoContent;
+#[derive(Clone, Debug, Default)]
+pub struct NoContent {
+	headers: HeaderMap
+}
 
 impl From<()> for NoContent {
 	fn from(_: ()) -> Self {
-		Self {}
+		Self::default()
 	}
 }
 
-impl ResourceResult for NoContent {
+impl NoContent {
+	/// Set a custom HTTP header. If a header with this name was set before, its value is being updated.
+	pub fn header<K: IntoHeaderName>(&mut self, name: K, value: HeaderValue) {
+		self.headers.insert(name, value);
+	}
+
+	/// Allow manipulating HTTP headers.
+	pub fn headers_mut(&mut self) -> &mut HeaderMap {
+		&mut self.headers
+	}
+}
+
+impl IntoResponse for NoContent {
 	// TODO this shouldn't be a serde_json::Error
 	type Err = serde_json::Error; // just for easier handling of `Result<NoContent, E>`
 
 	/// This will always be a _204 No Content_ together with an empty string.
 	fn into_response(self) -> Pin<Box<dyn Future<Output = Result<Response, Self::Err>> + Send>> {
-		future::ok(Response::no_content()).boxed()
+		future::ok(Response::no_content().with_headers(self.headers)).boxed()
 	}
 
 	fn accepted_types() -> Option<Vec<Mime>> {
@@ -55,7 +69,7 @@ impl ResourceResult for NoContent {
 }
 
 #[cfg(feature = "openapi")]
-impl ResourceResultSchema for NoContent {
+impl ResponseSchema for NoContent {
 	/// Returns the schema of the `()` type.
 	fn schema() -> OpenapiSchema {
 		<()>::schema()
@@ -67,7 +81,7 @@ impl ResourceResultSchema for NoContent {
 	}
 }
 
-impl<E> ResourceResult for Result<NoContent, E>
+impl<E> IntoResponse for Result<NoContent, E>
 where
 	E: Display + IntoResponseError<Err = serde_json::Error>
 {
@@ -86,12 +100,12 @@ where
 }
 
 #[cfg(feature = "openapi")]
-impl<E> ResourceResultSchema for Result<NoContent, E>
+impl<E> ResponseSchema for Result<NoContent, E>
 where
 	E: Display + IntoResponseError<Err = serde_json::Error>
 {
 	fn schema() -> OpenapiSchema {
-		<NoContent as ResourceResultSchema>::schema()
+		<NoContent as ResponseSchema>::schema()
 	}
 
 	#[cfg(feature = "openapi")]
@@ -104,7 +118,7 @@ where
 mod test {
 	use super::*;
 	use futures_executor::block_on;
-	use gotham::hyper::StatusCode;
+	use gotham::hyper::{header::ACCESS_CONTROL_ALLOW_ORIGIN, StatusCode};
 	use thiserror::Error;
 
 	#[derive(Debug, Default, Error)]
@@ -118,6 +132,9 @@ mod test {
 		assert_eq!(res.status, StatusCode::NO_CONTENT);
 		assert_eq!(res.mime, None);
 		assert_eq!(res.full_body().unwrap(), &[] as &[u8]);
+
+		#[cfg(feature = "openapi")]
+		assert_eq!(NoContent::default_status(), StatusCode::NO_CONTENT);
 	}
 
 	#[test]
@@ -127,5 +144,14 @@ mod test {
 		assert_eq!(res.status, StatusCode::NO_CONTENT);
 		assert_eq!(res.mime, None);
 		assert_eq!(res.full_body().unwrap(), &[] as &[u8]);
+	}
+
+	#[test]
+	fn no_content_custom_headers() {
+		let mut no_content = NoContent::default();
+		no_content.header(ACCESS_CONTROL_ALLOW_ORIGIN, HeaderValue::from_static("*"));
+		let res = block_on(no_content.into_response()).expect("didn't expect error response");
+		let cors = res.headers.get(ACCESS_CONTROL_ALLOW_ORIGIN);
+		assert_eq!(cors.map(|value| value.to_str().unwrap()), Some("*"));
 	}
 }